Change method of obtaining parent relation
There are cases where monkey-patching autobuild to build the parent
relation if it does not exist causes SystemStackError messages, most
noticably during unit testing when using factories to build the child
classes. It's simpler to simply override the association getter to build
the association if it does not already exist, and it gets rid of the
dangerous alias_method_chain.